[refactor] CustomLogger 위치 변경
