+++ BE/src/main/java/cuketmon/util/CustomLogger.java
@@ -1,4 +1,4 @@
-package cuketmon;
+package cuketmon.util;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
